pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "raptor"
version = "0.1.0"
license = { text = "BSD-3-Clause" }
authors = [
  { name="John Coleman", email="colemanjs@ornl.gov" },
  { name="Vamsi Subraveti", email="vamsi.r.subraveti@vanderbilt.edu" }
]
description = "A library to predict porosity in Laser Powder Bed Fusion (LPBF) builds."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: BSD-3-Clause License",
    "Operating System :: OS Independent",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "numpy",
    "numba",
    "PyYAML",
    "vtk",
    "scikit-image"
]

[project.scripts]
raptor = "raptor.cli:run"

[project.urls]
Homepage = "https://code.ornl.gov/ORNL-MDF/rapiddefectpredictor"
